feature: Stats added, now we can see the number of wins,loses and matches of one player. Aditionally i have made a small change in the Avatar, because whenever we tried to update the avatar this delete the default avatar.

+++ backend/django_backend/user/views.py
@@ -64,9 +64,10 @@ class AvatarViewSet(APIView):
         if(serializer.is_valid()):
             old_avatar = Avatar.objects.filter(user=user).first()
             if(old_avatar):
-                old_avatar.image.delete()
+                if old_avatar.image.name != 'default/default.jpg':
+                    old_avatar.image.delete() 
                 old_avatar.delete()
-            serializer.save(user=request.user)
+            serializer.save(user=user)
             return Response(serializer.data, status=status.HTTP_201_CREATED)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
